+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.0.0
 * Remove runmode from config path #32
+* Add AEM client parameter override support to integration tests
 
 ### 1.3.0
 * Add getCrxdeStatus operation